Refactored for a 'services' subfolder
